Move 'heading_size' out of 'label' property and into root hash
This brings the 'autocomplete' config closer in line with the
'select with search' config, which defines 'heading_size' in the
root hash rather than in the `label`. It appears
govuk_publishing_components is equally divided, so there is no
strong reason to go with one approach vs the other:
https://github.com/alphagov/govuk_publishing_components/issues/4641

Note that the 'select_with_search' component makes use of the
`GovukPublishingComponents::Presenters::SharedHelper` helper to
determine whether the passed 'heading_size' is valid. I did add
a test to 'autocomplete' with a view to adding that validation
to 'autocomplete', but the test passed without modification, so
it appears the component already avoids using invalid values and
thus there's no need for us to check for it in Whitehall:

```
  test "avoids passing heading size to label component if invalid value" do
    data = component_data
    data[:heading_size] = "sausages"
    assert_select ".govuk-label.govuk-label--sausages", { count: 0 }
    assert_select ".govuk-label", { count: 1 }
  end
```