Flatten 'label' property
There is no clear convention in govuk_publishing_components as to
whether a label should be a string or a hash.

Components with string labels:

- https://components.publishing.service.gov.uk/component-guide/copy_to_clipboard
- https://components.publishing.service.gov.uk/component-guide/password_input
- https://components.publishing.service.gov.uk/component-guide/select
- https://components.publishing.service.gov.uk/component-guide/search_with_autocomplete

Components with hash labels:

- https://components.publishing.service.gov.uk/component-guide/file_upload
- https://components.publishing.service.gov.uk/component-guide/character_count
- https://components.publishing.service.gov.uk/component-guide/input

Given we only now have a 'text' subproperty of 'label' for the
'autocomplete' component, the hash is a bit clunky to use and offers
no real benefit. So bringing it in line with 'select with search',
we're switching to a string, which will make it easier to consolidate
the components.