Merge details with existing details when updating
Before, when editing a draft, the details that a user gave overrode the
whole details object, meaning all embedded documents were removed.

This assigns only the attributes we’re interested in, as well as
merging the details hash from the form input, so anything that hasn’t
come from the form remains intact.

I’ve also had to update the `organisation_id=` method to set the
organisation to nil if `organisation_id` is an empty string.