Add support for multiple select in 'select_with_search' component
This is needed to deprecate the 'autocomplete' component.
Choices.js supports multiple select mode out of the box: we just
need to pass the `multiple: true` boolean to the Rails
`select_tag` helper (which adds the `multiple="multiple"` attribute
to the rendered `<select>` tag).

The resulting UI is a bit different to the 'autocomplete'
equivalent, but not so different that it should be detrimental
to our users. That said, in the next commit(s) we'll look at
replicating the existing 'autocomplete' UX.

Screenshots of 'multiple' support so far:

[Related guides](https://whitehall-admin.integration.publishing.service.gov.uk/government/admin/detailed-guides/new) dropdown. NB: there is a similar ([slow](https://govuk.zendesk.com/agent/tickets/5839628)) performance in both cases.

The landing page view has a similar affordance (we lose the 'dropdown' icon, but gain some placeholder text. Either one is relatively simple to add/remove if necessary). The search experience is identical too. What differs is the 'on choose' behaviour: 'autocomplete' exits the dropdown whereas 'select with search' keeps the dropdown open. The representation of selected options is different too ('autocomplete' shows the options below the input, and 'select with search' shows them inline).

| Before | After |
|-------|-------|
|![related_guides-closed-ac](https://github.com/user-attachments/assets/fd9e5beb-fc48-483a-9eb7-683c0530291c)|![related_guides-closed-sws](https://github.com/user-attachments/assets/9e0daee7-748d-4ed5-a725-04557981dbd2)|
| On page load | On page load |
|![related_guides-open-ac](https://github.com/user-attachments/assets/a1d8745f-45a3-42eb-b378-d8b2744d9f5c)|![Screenshot 2025-02-20 at 14 54 38](https://github.com/user-attachments/assets/54b647bf-f8b6-4647-860e-6e57a98bdea0)|
| On focus | On focus |
|![related_guides-selected-ac](https://github.com/user-attachments/assets/f6821fbb-2520-459a-8cf1-c6e0d2725760)|![Screenshot 2025-02-20 at 14 51 59](https://github.com/user-attachments/assets/7d3e2c1d-157d-4596-8895-49fdcf6ddfbc)|
| On choice made | On choice made |
|As above |![related_guides-selected-sws](https://github.com/user-attachments/assets/2a56331d-6381-4398-ade0-e4eb8ed740ca)|
| On blur | On blur |

Trello: https://trello.com/c/jBjHrNou/3492-replicate-ux-of-multiple-autocomplete-into-selectwithsearch-component