Change how TaggableContentHelper caches queries
The query caching (added in #1712 and #1728) is important to
retain as, according to the linked PRs, it causes considerable
savings in page load times for the affected pages.

However, in one of the next commits we'll be changing the structure
of the returned data to include not only the `text` and `value` of
each option, but also its `selected` status. We don't want to
cache the `selected` status as that would mean for example we're
showing the same "Lead Organisation" selected item across all
five "Lead organisations" dropdowns on the "New document" page.

So, we're keeping the caching, but we're just going to cache the
expensive bit - the database result. The `map` calls are very
quick in comparison and don't need to be cached.