Convert 'select with search' error properties into error_items
This was needlessly different to the 'autocomplete' implementation
which takes an array of errors, preventing consolidation of the
components.

The 'select with search' component was just following what was
defined in the 'select' component:
https://components.publishing.service.gov.uk/component-guide/select#with_error

It 'delegated' the error logic to that component, which is quite
a nice pattern.

However, the prevailing opinion seems to be that the `error_items`
array is the way to go. See elsewhere in Whitehall, e.g.
https://github.com/alphagov/whitehall/blob/6b00ce8873698a556ff3db44acf0a45c137f2e1c/app/views/admin/attachments/_form.html.erb#L12
See also the commit message that introduces the `errors_for_input`
method, which adds a special method just to support `error_message`
for the `input` component, whereas "All other [components] require
an array of hashes".

NB: `error_id` being a specifyable thing was weird: it didn't
appear to be being manually set in application code anywhere.
By moving to `error_items`, we've removed that unnecessary bloat.

Note the mass switch from `errors_for_input` (which squashes to a string)
to `errors_for` (which converts to an array of hashes as expected)
The `errors_for_input` class should only really be used by the
`input` component (as explained in the commit linked above).