Remove unused `is_page_heading` property
This is a case of YAGNI. It's supported by the 'select' component
(which uses it in one place in Whitehall: https://github.com/alphagov/whitehall/blob/514ba658e64b3dd3b5bb12a607e47f3f0d80422c/app/views/admin/worldwide_organisations_main_offices/show.erb#L32)
but it isn't used by any select_with_search component. We can
always add it back in later if we need it.

Removing it simplifies the proposition between 'autocomplete' and
'select with search' (i.e. we don't need to port over the
`is_page_heading` behaviour to the 'autocomplete' component if we
were to decide to switch over to that component exclusively).