Replace all calls to 'autocomplete' with 'select_with_search'
This commit switches over all of our 'autocomplete' usage to use
the 'select with search' component instead. Thanks to the API
unification PR that precedes this one, there's nothing else we
need to do to get this working, in theory...!

In _practice_, the changes caused a few failing tests. It appears
the ID of the rendered `<select>` is ever so slightly different
between the two components. See #8237 and #8716 for the first
references to the weird 'edition_supporting_organisation_ids_' ID.
The ID is now 'edition_supporting_organisation_ids' (no trailing
underscore).