Remove 'autocomplete' component
Now that all of the 'autocomplete' component usage has been
changed to call 'select with search' instead, we can delete the
'autocomplete' component ERB partial, tests and component YAML
docs.

Unfortunately we're unable to delete the JS and CSS, and the
accessible-autocomplete-multiselect, since these are used by
miller-columns-element. I've renamed the files accordingly, to
make it clear that this is the only place they're used now.
Once accessible-autocomplete-multiselect is removed from all of
the apps apart from Whitehall, we should strongly consider
consolidating the JS/component into miller-columns-element so
that it has search functionality supported right out of the box.