update pension rate examples to expect full copy
previously we had the enums as 'weekly',
'monthly' etc, but we now want to use this
field as if it was the exact copy content
designers would be pasting into documents.

So we no longer call `humanize` on the enum
because it should already be the humanized
version. I think this is ok to do for now, as we do
not have any other enums in CBM afaik, and if it
causes problems we will need to revisit.