Use Fulltext matching to search by keyword
This should be faster than using regular expressions. The tests don’t
work out of the box, because they’re run in transactions, so the indexes
don’t get built. To get around this, we need to disable transactions for
these tests and make sure DatabaseCleaner wipes that database after each
test run.