Add index to the title, details and instructions
This allows us to search the fields using MySQL’s fulltext search
functions. Because we can’t index JSON fields using fulltext, we have
to create a virtual column which casts the JSON as text. I tried some
MySQL-fu to try and extract all the JSON values, but we can’t do it
recursively. This should be enough anyway, as it’s what we’re only
searching across the JSON at the moment.