Fix bug resulting in failure to publish worldwide orgs
When a new edition of a worldwide org was created, the main office association was still pointing to an office belonging to the previous edition. This was causing errors when attempting to publish the new edition because the edition ended up with a duplicate contact link.

We have to save the new edition without validating after setting the main office, because the edition is invalid due to a missing change note.