Swap 'Publication type' `select_with_search` for `autocomplete`
This dropdown has just 20 options, so doesn't particularly need
search functionality. It does have a nice grouping of publication
types as to whether they're commonly used, less commonly used, or
'use discouraged'.

I tried simply swapping 'select with search' out for 'select', but
the govuk_publishing_components [select component](https://components.publishing.service.gov.uk/component-guide/select)
doesn't support grouped options. Ideally we would port the
existing 'select with search' grouped options styles over to the
'select' component and then release a new version of
'govuk_publishing_components', but there's little evidence it
would be used outside of just two cases in Whitehall, so it may
be a difficult case to make, and in the meantime, we want to
retire the 'select with search' component.

I've therefore swapped this over to use the 'autocomplete'
component instead. We lose the nice 'grouping' of options, but
I've added a prefix to the 'less common' and 'use discouraged'
publication types to keep some degree of visual separation.
NB: I've reworded the latter to 'deprecated' as it works better
as a prefix, and because if the use of these types is discouraged
then they are by definition not well supported.