Hide details if there is no change history
Sometimes change history can be present, but doesn’t have items within
it. To get around this, we move the logic to the component and only
render the details if there are any details rendered.

Additionally, I’ve changed the version factory to build, rather than
create so we don’t have to create versions in the database
unneccessarily