Simplify link check report association with editions
The polymorphic association was added in November 2017 with no
explanation: https://github.com/alphagov/whitehall/pull/3557/commits/03c47344a42243f96990559e33125f73edc4d00b

Since then, there doesn't seem to have been a single link check
report which is associated with anything other than an Edition:

```
$ LinkCheckerApiReport.distinct.pluck(:link_reportable_type)
=> ["Edition"]
```

Therefore let's remove the obtuse reference and make it clear
that a link check report is associated with an edition (not a
'linkable'). This should also cut down on quite a lot of DB space,
since we're currently storing the word "Edition" on every single
link check report (almost 12 million rows in the DB).

We'll deploy this and run the migration, before removing the old
fields.