Add data migration to fix timestamps of previous data migrations
See previous commit. Timestamps were incorrectly formatted, which
ended up being copied into the database too. In order to avoid
re-running the same migrations that have already been run, we
need to update the timestamps to the new (correct) timestamp that
has been applied to their filenames. We also need to do this
correction in a data migration that happens 'before' the already
run data migrations, otherwise the rename won't happen in time.
I've therefore set a fake timestamp of 1st July 2024 so that this
database correction happens before the first offending migration.

Note that I also had to change the implementation of the
`DataMigrator.run` method, so that it fetches a 'fresh' list of
"due" data migrations each time one has completed. That allows us
to rename the data migrations that follow the one I've introduced
in this commit, and avoid running them twice.
I considered adding a test for this but it's such a one-off that
the existing tests are more than OK.