Don't parallelise tests marked with `RUN_IN_ISOLATION`
This allows us to add a `# RUN_IN_ISOLATION` magic comment to any
test that is sensitive to being run in parallel, which so far is
just the `ScheduledPublishingWorkerTest`, because it queries the
Sidekiq::ScheduledSet directly, and that set can have its contents
altered by other tests at runtime.

Test files with the magic comment are put onto their own node in
GitHub Actions, and automatically configures the parallelize gem
to run tests sequentially on that node, so all of our flaky tests
can be run one after the other, at a slower speed but a greater
reliability.

I really didn't want to do this. I tried switching up this test
to use Sidekiq's "inline mode" for isolation, which forces jobs to
execute immediately instead of being queued, removing race
conditions. It's the approach we've already been applying
elsewhere in Whitehall:
https://github.com/alphagov/whitehall/blob/d045ee80abfa1aad6750b779132b4baf3ca03adf/test/unit/lib/whitehall/publishing_api_test.rb#L98

Unfortunately it’s not as simple as setting the
`scheduled_publishing_worker_test` into inline mode, as it’s some
other test(s) that are polluting the ScheduledSet that are causing
it to fail. So we’d have to go and fix all other tests to make
this one test file pass consistently. That could quickly turn into
whackamole...!

I then looked into [forcing Sidekiq to run inline globally](https://github.com/alphagov/whitehall/commit/a5a0589d89a05c8903592ffc75fea4cb4ce6c494),
but that didn't appear to make any difference (nor did it allow me to
[delete a whole bunch of 'Sidekiq::Testing.inline!' calls](https://github.com/alphagov/whitehall/commit/7cb6c5cb18a564fe3aef357411ed55b4f98e3c1c?w=1)
as I'd hoped)