Use singular `link_check_report`, remove references to the plural
This was mostly straightforward, though the change at
`app/views/admin/editions/show/_sidebar.html.erb` where I had
to swap out `@edition.link_check_report.build` for
`LinkCheckerApiReport.new(edition: @edition)` took a bit of
thinking. It looks like we always need to be sending a valid
`report` object to the template, otherwise it fails on all
these method calls:
https://github.com/alphagov/whitehall/blob/a49f2c3d686dcba096e481597fe3c80d7a59009b/app/views/admin/link_check_reports/_link_check_report.html.erb#L2

The `if report.new_record?` conditional in that template seems
to check for if the object we've passed has been saved to the DB.
If it hasn't, it's a "new record" and thus we see the 'first time
message' which is "Check this document for broken links. The report
will take a few moments to complete". This differs from repeat
visits where the message is simply "Check again". So what the
template is expecting is a LinkCheckReport that lives in memory
only, as a sort of placeholder.