Update existing - rather than add new - link check reports
In the next commit, we'll see the pay-off of this work, which is
that every edition will only have one `link_check_report` and
we'll no longer have to deal with arrays of link check reports
(which have been an unnecessary complexity, given we never do
anything with the old link check reports).

But before we can remove `link_check_reports`, we need to do the
work to create-or-update an existing link check report, so that
editions only ever have 0 or 1 link check reports associated with
them. That's what this commit is all about.
(Removing legacy link check reports will be covered in another
commit).

Note that we've removed unnecessary calls to `capture_subprocess_io`.
[Its own documentation](https://www.rubydoc.info/gems/minitest/5.10.3/Minitest%2FAssertions:capture_subprocess_io)
says "this method is approximately 10x slower than #capture_io so
only use it when you need to test the output of a subprocess".
We don't test the output of the sub-process. I imagine we only
include this line because we want to suppress the debugger output
from the link checker. We can find another way of doing this if it
becomes annoying.

Note also the updated feature test: we're now overwriting the
existing link check report rather than creating a new one each time,
so we should just use the same ID throughout the test.