Remove unused edition policy code
The last record in the edition_policies table was created in Decemeber 2019. I haven't been able to trace exactly what it was for, but it seems to have been a way to link some edition types to a policy document (which may or may not have been published by Whitehall as the policy document was referred to by content ID).

There seems to have been a feature where documents could be marked as relevant to local government which used the edition policy relationship in some way, but that has been replaced by a taxon (see https://www.gov.uk/regional-and-local-government/local-government).

All in all it seems this code is safe to remove