Surface fixes for link validation errors
Currently fixes for any link validation errors that prevents publishing and scheduling are silent. The user only sees a high level message that there was something wrong with the links and not how to go about fixing it. If we surface the fixes we may receive less support tasks to manage.