Don't retry jobs that fail with `GdsApi::HTTPPayloadTooLarge`
We're seeing a lot of republish job failures off the back of a
recent bulk republish, whereby the documents in question are just
far too large for the publishing platform to handle (too many
attachments etc):
https://govuk.sentry.io/issues/6332398641/?environment=production&project=202259&referrer=issue-stream&statsPeriod=1h&stream_index=0

It's pointless retrying these, as the payload size won't change in
between retries, and therefore it's always going to fail. We
should instead log the exception, and kill the job.