Change data structure as requested by PA
PA testing has revealed that the data structure which is required
is not the same as what was provided in the Trello card. Specifically,
the data shown in the Trello card description should be included as
an `event_name` property in a wrapper object that also has an
`event: 'event_data'` property.