Another fixup
Removed unnecessary call to 'capture_subprocess_io'. The docs say
'This method is approximately 10x slower than #capture_io so only use it when you need to test the output of a subprocess.'
We don't test the output of the sub-process. I imagine we only
include this line because we want to suppress the debugger output
from the link checker. We can find another way of doing this.

(TODO: let's remove the only other instance of capture_subprocess_io
too)