WIP: fixed the issue
I'd made an incorrect assumption. I thought that
CreateFromBatchReport would be used to create a new link check
report if none exists already, and UpdateFromBatchReport would
be used to overwrite existing link check reports. And therefore
for new editions, we'd call the former, and for existing editions
we'd call the latter.

In fact, we ALWAYS want to call CreateFromBatchReport, even though
we're now replacing the report. The 'UpdateFromBatchReport' method
is used _to overwrite the just-generated report_.

I.e. CreateFromBatchReport creates an 'in progress' report, and
'UpdateFromBatchReport' updates it when it receives the callback
from Link Checker API.