Remove use of 'capture_subprocess_io'
These were getting in the way of my `require 'pry'; binding.pry`
calls. They're also unnecessary. According to
[its own documentation](https://www.rubydoc.info/gems/minitest/5.10.3/Minitest%2FAssertions:capture_subprocess_io),
"this method is approximately 10x slower than #capture_io so
only use it when you need to test the output of a subprocess".

We don't test the output of the sub-process. I imagine we only
include this line because we want to suppress the debugger output
from the link checker. We can find another way of doing this if it
becomes annoying.