Simplify `CreateFromBatchReport` class
In #3675, the class was amended to replace a link check report if
one already exists with the same batch ID. This is a bit confusing
given the name of the class was never changed:
`CreateFromBatchReport` implies there's only ever a create
operation.

The reason provided was this:

> We're coming across a problem whereby the data sync is causing a
> mismatch in the batch IDs stored in Whitehall and the batch IDs
> stored in the Link Checker API. This is because the databases
> don't get copied across at the same time.
>
> We thought the best solution to this would be to simply replace
> any existing link report that exists with the same batch ID.

In hindsight, the approach is somewhat flawed. Whilst it prevents
`ActiveRecord::RecordNotUnique` exceptions (because it overwrites
a link check report, rather than creating a new one with the same
batch ID), it hides a potentially more serious problem: that it's
overwriting the link check report of a totally unrelated edition!

We've therefore decided to change the order in which the data
sync runs, so that Whitehall's data sync completes before Link
Checker API's data sync begins. That way, Link Checker API should
only ever be 'ahead' of Whitehall (e.g. Link Checker API's max
batch ID may be 1000, whereas Whitehall's latest batch ID may be
950. The next time Whitehall requests a new link check report,
Link Checker API returns a report with batch ID of 1001, which
is fine as it won't conflict with any existing reports in
Whitehall).

See https://github.com/alphagov/govuk-helm-charts/pull/3058