Rename link check report creation/updating methods
These methods were somewhat generically named, and it made it hard
to understand. "Create from batch report" - does that create a new
report or, as may have been inferred from reading the code removed
in the last commit, does it also include updating existing reports?

"Update from batch report" - update what? Does it create a new one?
Etc.

The new method names should be a bit clearer. It's clear we're
creating a report with an "in_progress" state in one, and then
updating an existing report and marking it as "completed" in
another.