Replace - rather than append - link check reports
It's a shame we had to duplicate the `delete_previous_reports`
method for both `CreateFromBatchReport` and `CreateNoopBatchReport`,
but it's really very minimal code and didn't seem to warrant
further abstraction.

I did consider whether it's better to find and update an existing
row in the DB, or whether deleting old ones and creating a new
one is the way forward. I've opted for the latter out of simplicity
(since there isn't always going to be an existing row to update)
but don't feel too strongly about it either way.

The `CreateFromBatchReport` deletion behaviour is covered by the
existing tests in `check_organisation_links_worker_test`, touched
in this commit. (I've verified that the tests fail if the deletion
logic is removed).