Send GA4 data upon changes to select_with_search
Introduces a new analytics GA4SelectWithSearchSetup module that sends
data to GA4 in response to the custom `addItem` and `removeItem` events
provided by choices.js. The requirements provided by Performance Analysts
are for:

* both individual select and multi-select instances of select_with_search
  to send data to GA4 when a user makes a selection (this has been achieved
  through an `addItem` event handler)
* only multi-select instances to send data in response to `removeItem` events.
  The relevant `removeItem` handler is therefore only attached to those instances
  with the `multiple` attribute

Following the implementation of other analytics modules, a new module is added
to the HTML (`ga4-select-with-search-setup`) to allow for control of when this
tracking is enabled.