Send GA4 data in response to 'change' events on HTML select elements
Introduces a new analytics GA4SelectSetup module that sends data to GA4
in response to `change` events on only those select elements that are
not part of a `select_with_search` component (See #9983 for details of
how the `select_with_search` components are tracked using custom events
provided by choices.js).

The `select_with_search` components have been excluded through the selector
used to build the nodeList:

"select:not([data-module~='select-with-search'] select)"

This instructs the DOM to return all select elements that do not have
an ancestor decorated with the 'select-with-search' data-module. This
approach is necessary to avoid multiple submissions in response to a
single change in `select_with_search`.

Following the implementation of other analytics modules, a new module is added
to the HTML (`ga4-select-setup`) to allow for control of when this
tracking is enabled.