Send GA4 data in response to 'change' events on HTML select elements
Introduces a new analytics GA4SelectSetup module that sends data to GA4
in response to `change` events on only those select elements that are
not part of a `select_with_search` component (See #9983 for details of
how the `select_with_search` components are tracked using custom events
provided by choices.js).

Following the implementation of other analytics modules, a new module is added
to the HTML (`ga4-select-setup`) to allow for control of when this
tracking is enabled.

By default `event_data.type` is filled by a data attribute specified on
the moduleElement that the select element is a child of. If the select
element specifies a `ga4DocumentType` then this will be used instead.

By default `event_data.section` is filled by the label of the select
element. If the select element specifies a `ga4Section` then this will
be used instead.