Create RemoveDangerousLinksWorker on link check report update
This worker is only instantiated from the `mark_report_as_completed`
method, which is only called in the `LinkCheckerApiController`
callback once Link Checker API returns a completed batch report.

We intend to hook into this callback and automatically remove
any dangerous links from live editions, and republish said
editions. The implementation for that removal and republish will
be in the next commit.