Automatically remove dangerous links from published editions
If we're informed of a domain that is no longer safe, but is
referenced across a number of our pages, then it is not feasible
for our content designers to step in and manually fix them all in
a timely manner. Instead, we need an automated approach. This
commit introduces a blunt tool to find and replace all references
to the dangerous domain with a `#link-removed` value.

Whilst ideally we'd have a content designer address these on a
case by case basis - finding an alternative URL to link to, for
example, or removing the link but adding some historical context -
we're not able to do that in an automated way. The alternative is
to either:

1. remove the link and link text altogether, which could lead
   to a confusing 'absence' of content
2. remove the link but keep the link text, which could be odd if
   the link text reads something like "click here"
3. replace the link with a `#`, so that it's still a link but
   doesn't do anything.
4. replace the link with an anchor tag that leads nowhere, i.e.
   `#link-removed`, so that it's still a link, it doesn't do
   anything, and it's clearer to the user and to the publisher
   what has happened.

I've opted for option 4. I've done so with minimal HTML / GovSpeak
wrangling, as I can't envisage a situation where we'd even want to
display the name of a dangerous domain, so finding-and-replacing
only `<a>` link tags seemed an unnecessary complexity when we can
simply find and replace all instances of the string.

So how does it work architecturally?

This new `RemoveDangerousLinksWorker`, triggered via the Link
Checker API batch report callback only if a dangerous link was
detected, does the following:

1. Validates that the edition is 'published'. There's little point
   in fixing up superseded editions, and 'draft' editions haven't
   yet been published, so we're better off surfacing dangerous
   link warnings in the UI for the publisher to resolve it
   properly.
2. Validates that there is no existing 'draft' edition. We don't
   want to risk overwriting the publisher's draft with an auto
   generated draft edition that is identical to the live edition
   but with dangerous links removed.
3. Validates that there are indeed 'danger' links in the link check
   report, that need removing.
4. Creates a draft edition, removes the dangerous links, and force
   publishes the edition as a minor change, with an internal note
   explaining what has happened.

Since https://github.com/alphagov/whitehall/pull/10077, we should
be refreshing every link check report for every published edition
on GOV.UK roughly every 7 days, meaning we don't have to wait too
long for all dangerous links to be systematically removed.
Dangerous links will be driven by the configuration of 'suspicious
domains', as will be documented in https://github.com/alphagov/link-checker-api/pull/988.

There is room for improvement, but it can come later. Suggestions:

1. Check that the body of the edition changes after calling
   `remove_danger_links!`, before bothering to create a draft.
2. Better handle the "published with draft" state.
3. Add support for "scheduled" states.

Trello: https://trello.com/c/tmnht4P1