Make sure multiple select with search sends a nil value when all selections are removed
HTTP form data is unable to represent empty arrays. If a select has a name like "test[]" and no selected value then the name of the select will be dropped from the form data when the result is submitted. However, by creating a hidden input with the same name we can make sure that the form data will look like 'test[0]=""' instead of 'test=[]'. If the select has a value, then that will override the nil value.