Fix Audit Log
Whilst we were creating the draft edition as "Scheduled Publishing
Robot", our audit gem requires that we wrap all method calls in an
'acting_as' block.

Screenshot of the attribution working:

![attributed to Scheduled Publishing Robot](https://github.com/user-attachments/assets/1aa4b5d5-7723-4311-bda4-0a98e596e80a)

We can see in the screenshots in #10081 that without this commit,
attributions are linked to "User (removed)".

Trello: https://trello.com/c/tmnht4P1