Temporarily avoid creating RemoveDangerousLinksWorker
This hasn't run on Production yet, but has been tested locally and
on Integration and Staging. It seems to be working well, but there
is a potential logic loop that could cause multiple republishes
(see #10091), and locally we're seeing more 'editorial remarks'
being created than we should.

We're therefore commenting out the bit of code that causes a
republish, to allow time to investigate further and gain a bit
more confidence before releasing to Production.

Trello: https://trello.com/b/0n3p84mY