Remove the "allow_new_report" logic
Now that we're surfacing the "Last checked" timestamp, there's no
reason to keep this unnecessary complexity around showing/hiding
the "Check again" button. Let's just let publishers check as often
as they want. (Which they can already do, they just need to refresh
the page first).

It looks as though the decision to conditionally show a "Check
again" button was made in https://github.com/alphagov/whitehall/pull/7325,
allegedly to retain parity with the pre-Design-System page. But as
far as I can tell, this was never the case since the link check
report feature was first added in https://github.com/alphagov/whitehall/pull/1573.

The only case that seems to be made for hiding the "Check again"
button is when there is a report already in progress, and that
is something that is retained in this PR.