Add 'warning' state to link check report
Hypothesis: the "broken links" error message and styling that we
give to users is unhelpful and unactionable for 'warnings'. For
example, the "Page blocks robots" warning indicates that Link
Checker API has been unable to check the link, but the link may
very well be valid and there may very well be nothing for the
publisher to do.

In this case, marking the report as containing "broken links" is
incorrect and unhelpful. Instead, it's better to render a 'warning'
kind of UI so that we can warn publishers that they'll need to
manually check the affected link(s).