Remove Attachment deletion rake task
In light of our current understanding of asset management between Whitehall and Asset Manager, this rake task is no longer necessary.
If there is a requirement to remove an HTML attachment, we use `redirect_html_attachments` rake task.
If there is a requirement to remove a file attachment, there are instructions to remove them directly from asset manager. The deletion state of attachments in Whitehall is more about the link between Attachments and its corresponding Edition, which means users should be able to break that link within the UI.