show content block updates whatever state host content is in
If a Document has an embedded content block, we
want to show changes to that content block in the
document's history, if it is any state except
`unpublished`. `unpublished` looks like the only
final state when a Document will have no further
changes, from looking at `workflow.rb`,
`withdrawn` can be `unwithdrawn`.