Add JavaScript to populate ga4-index-section client-side
Introduces a JavaScript module that adds the data attribute
client-side, rather than the previous approach of doing this
in Rails. This comes from a suggestion in the PR[^1] arising
from the complexities of knowing the index at the time HTML
is produced by Rails.

The core part of the module is the CSS selector that is used:

`select, input:not([data-module~="select-with-search"] input)`

This builds a nodeList comprising all selects and those inputs
which are not part of a `select-with-search` component. The
CSS selector can be amended in future to reflect PA needs.

[^1]: See https://github.com/alphagov/whitehall/pull/10069#discussion_r2022600260