Reindex parent organisations when child org added/destroyed
We've had several reports of Signon users having difficulty
accessing Signon because they're a part of a child organisation
and - unexpectedly - org editors of the parent organisation have
been unable to administrate their accounts.

On inspection, the orgs in Signon are out of sync with the orgs
in Whitehall. In the most recent example, the 'skills-england' org
has no parent in Signon.

Looking at skills-england in Organisations API, we can see it does
have a parent of 'department-for-education':
https://www.gov.uk/api/organisations/skills-england

The corresponding entry for 'department-for-education', however,
lacks a 'child_organisations' relationship with 'skills-england':
https://www.gov.uk/api/organisations/department-for-education

Since [Organisations API is populated by Search API](https://docs.publishing.service.gov.uk/manual/organisations-api.html#populating-the-organisations-api),
we need to dig into Search API to find the issue. Sure enough, the
corresponding Search API results for each org match what we're
seeing in Organisations API:
- https://www.gov.uk/api/search.json?filter_link=/government/organisations/skills-england
- https://www.gov.uk/api/search.json?filter_link=/government/organisations/department-for-education

Looking in Whitehall, we had no special logic handling the indexing
of parent organisations. We know we have to manually update orgs
in Search API after save/destroy, but this was only being done
on a per-organisation basis, so was never updating the parent/child
relationship in Search API.

I've added some additional callbacks to update all parent organisations
in Search API whenever a corresponding child org is created or
destroyed.

PS: some 'reload' calls were required to fix a couple of tests
that started failing as a result of this. I'm not entirely sure
why, but I don't see any harm in it, since in production we don't
tend to create and use an org and its attributes so procedurally.

Trello: https://trello.com/c/oPEsW9mH/3605-investigate-mismatch-between-whitehall-organisations-and-search-api